Python3/Tornado/apps/pg/PG_AdminREST/requirements.txt
django==3.0.5
Pygments==2.6.1
pika==1.1.0
requests==2.23.0
django-redis==4.12.1
redis==3.5.3
mnemonic==0.19
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.4.0
drfdocs==0.0.11
mysqlclient==1.4.6
django-filter==2.3.0
ed25519==1.5
cryptography==3.0
coreapi==2.3.3




